casted some variables and added an extern to clean up some warning messages...
git-svn-id: svn://fly.thruhere.net/forsaken@352 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/Primary.c
@@ -6752,8 +6752,8 @@ void DispPowerLevel( void )
 	int16	Steps, Count;
 	int16	x, y;
 
-	x = (d3dappi.szClient.cx / 2 ) - ( 16 * 4 );
-	y = d3dappi.szClient.cy - ( d3dappi.szClient.cy / 5 ) + 6;
+	x = ((int16)d3dappi.szClient.cx / 2 ) - ( 16 * 4 );
+	y = (int16)d3dappi.szClient.cy - ( (int16)d3dappi.szClient.cy / 5 ) + 6;
 
 	switch( Ships[ WhoIAm ].Primary )
 	{
